requirement.txt
Django>=5.2
djangorestframework>=3.15
django-cors-headers>=4.4
psycopg>=3.2
cryptography>=42.0
djangorestframework-simplejwt>=5.3
google-auth>=2.0